requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
python-multipart

# Database
sqlalchemy
alembic
asyncpg
psycopg2-binary

# Redis
redis[hiredis]

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt

# Email
resend

# HTTP client
httpx
aiohttp

# Development
pytest
pytest-asyncio
black
flake8
mypy
ipython

# Monitoring
structlog
python-json-logger

# Utilities
tenacity
arrow 